added & integrated MapData class
MapData offers a more convenient way to work with the arrays of values that make up map data to be sent to a client.
The class handles its own packing and unpacking.
It has helper methods that allow it to emulate a singe, hashed array, instead of seperate arrays

Example usages:
creation
	$simple_chunk = MapData.new(Units::BlockLength.new(16),Units::BlockLength.new(128),Units::BlockLength.new(16))
setting a single block's properties by array index
	$simple_chunk[77]={:type=>33,:metadata=>0,:block_light=>3,:sky_light=>15}
you can pass a hash of only the types you want to change
	$simple_chunk[78]={:type=>7}
	$simple_chunk[80]={:sky_light=>100}
you can use a triplet of x,y,z values as an alternate array index
	$simple_chunk[5,2,90] = {:type=>8,:sky_light=>10}