Added units for time, fixed inventory bug, fixed time packet
because ruby doesn't have a convenient 'unpack' for long integers in network order, for now we'll just treat the 'time' value for the TimeUpdate packet as a high int and low int.

added units and convenience methods for time related matters - MinecraftHours and MilliHours
minecraft's day cycle has about 24000 increments of the time value per cycle. So, I'm going to call each increment a millihour, with 1000 millihours per MinecraftHour, and 24 MinecraftHours in a MinecraftDay

InventoryUpdate packet now reads and writes correctly. I may go back and re-add the IO.read_network_* functions, for convenience & readability in this method