Refactored main file into seperate server/model classes
The main file, rubymine.rb, will run the server thread. this will be concerned with connecting clients, the handshake portion of the protocol, and reading packets. it will pass of the rules and heavy logic to the WorldModel class

the WorldModel class will mantain the state of the map, the positions and other properties of the players, and will interperet Packet objects to alter the state of the world accordingly, and to update players through their to_io methods.

Minor bugfixes or feature adds:
*map data now invalidates old cached value when it is updated
*fancy pretty print for minecraft_units
*more elegant metaprogramming for numeric convenience methods for units, based on @pp_name
*DestroyEntity and EntityInitialize packets now correctly list entity_id's type as :int, not :id
*Players now track their own inventory